+++ stereo_vision/new_function/core/DIC_core.c
@@ -7,7 +7,7 @@
 #include "include/PSO_comm.h"
 
 extern int interp_safe_margin();
-extern float bilinear(float *data, int width, int height, float x, float y);
+extern double bilinear(double *data, int width, int height, double x, double y);
 
 void run_PSO(struct SYS_INFO *info) {
 	// create PSO_context and put DIC info in it
@@ -21,29 +21,29 @@ void run_PSO(struct SYS_INFO *info) {
     }
 }
 
-float get_subset_mean(const DIC_subset_info *subset) {
+double get_subset_mean(const DIC_subset_info *subset) {
 	if (!subset) return NAN;
-    float sum = 0.0f;
+    double sum = 0.0f;
     for (int i = 0; i < subset->total_pixels; i++) {
-        sum += (float)subset->subset_data[i];
+        sum += (double)subset->subset_data[i];
     }
     return sum/(subset->total_pixels);
 }
 
 static void get_ref_subset_data(struct SYS_INFO *info) {
 	struct DIC_ZNCC_context *zncc_ctx = info->pso_ctx.priv;
-	int *pos_ptr 				= info->dic_ctx.img_ref_pt_pos;
-	float *ref_img_data 		= info->pso_ctx.img_info.ref_data;
+	double *pos_ptr 				= info->dic_ctx.img_ref_pt_pos;
+	double *ref_img_data 		= info->pso_ctx.img_info.ref_data;
 	int subset_side_len			= info->dic_ctx.config.subset_side_len;
 	int img_width 				= info->pso_ctx.img_info.width;
 	int img_height 				= info->pso_ctx.img_info.height;
-	float img_pt_y_shift 		= pos_ptr[0] - (subset_side_len - 1) / 2.0f;
-	float img_pt_x_shift 		= pos_ptr[1] - (subset_side_len - 1) / 2.0f;
-	float *ref_subset_data 		= zncc_ctx->ref_subset_info.subset_data;
+	double img_pt_y_shift 		= pos_ptr[0] - (subset_side_len - 1) / 2.0f;
+	double img_pt_x_shift 		= pos_ptr[1] - (subset_side_len - 1) / 2.0f;
+	double *ref_subset_data 		= zncc_ctx->ref_subset_info.subset_data;
 	for (int row = 0; row < subset_side_len; row++) {
 		for (int col = 0; col < subset_side_len; col++) {
-			float img_pt_y_shift_row = img_pt_y_shift + row;
-			float img_pt_x_shift_col = img_pt_x_shift + col;
+			double img_pt_y_shift_row = img_pt_y_shift + row;
+			double img_pt_x_shift_col = img_pt_x_shift + col;
 			int pixel_idx = row * subset_side_len + col;
 			ref_subset_data[pixel_idx] = bilinear(ref_img_data, img_width, img_height,
 												img_pt_x_shift_col, img_pt_y_shift_row);
@@ -66,7 +66,7 @@ void ZNCC_ctx_init(struct SYS_INFO *info) {
 }
 
 int ZNCC_subset_alloc(struct DIC_ZNCC_context *zncc_ctx) {
-	int subset_memory_size = square(zncc_ctx->cur_subset_info.side_len) * sizeof(float);
+	int subset_memory_size = square(zncc_ctx->cur_subset_info.side_len) * sizeof(double);
 	zncc_ctx->cur_subset_info.subset_data = malloc(subset_memory_size);
 	zncc_ctx->ref_subset_info.subset_data = malloc(subset_memory_size);
 	if (!zncc_ctx->cur_subset_info.subset_data || !zncc_ctx->ref_subset_info.subset_data) { 
@@ -102,27 +102,27 @@ int ZNCC_init(struct SYS_INFO *info) {
 	return 0;
 }
 
-float ZNCC_cost_function(struct PSO_context *ctx) {
+double ZNCC_cost_function(struct PSO_context *ctx) {
 	// SYS_DBG("ZNCC_cost_function\n");
 	if (ctx == NULL) return -10.0f; // result range: -1 ~ +1
 	struct DIC_ZNCC_context *zncc_ctx = (struct DIC_ZNCC_context *)ctx->priv;
-	float rel_pso_pt_y 				= zncc_ctx->subset_pt_cur_pos[0]; 	// Pi_x, Pi_y relative corrected coordinate
-	float rel_pso_pt_x 				= zncc_ctx->subset_pt_cur_pos[1]; 	// Pi_x, Pi_y relative corrected coordinate
-	float img_ref_pt_y 				= zncc_ctx->img_pt_ref_pos[0]; 		// start point (fixed in whole PSO process)
-	float img_ref_pt_x 				= zncc_ctx->img_pt_ref_pos[1]; 		// start point (fixed in whole PSO process)
+	double rel_pso_pt_y 				= zncc_ctx->subset_pt_cur_pos[0]; 	// Pi_x, Pi_y relative corrected coordinate
+	double rel_pso_pt_x 				= zncc_ctx->subset_pt_cur_pos[1]; 	// Pi_x, Pi_y relative corrected coordinate
+	double img_ref_pt_y 				= zncc_ctx->img_pt_ref_pos[0]; 		// start point (fixed in whole PSO process)
+	double img_ref_pt_x 				= zncc_ctx->img_pt_ref_pos[1]; 		// start point (fixed in whole PSO process)
 
 	int img_width 					= ctx->img_info.width;
 	int img_height 					= ctx->img_info.height;
-	float *cur_img_data 			= ctx->img_info.cur_data;
+	double *cur_img_data 			= ctx->img_info.cur_data;
 	int subset_side_len 			= zncc_ctx->ref_subset_info.side_len;
-	float *ref_subset_data 			= zncc_ctx->ref_subset_info.subset_data;
-	float *cur_subset_data 			= zncc_ctx->cur_subset_info.subset_data;
+	double *ref_subset_data 			= zncc_ctx->ref_subset_info.subset_data;
+	double *cur_subset_data 			= zncc_ctx->cur_subset_info.subset_data;
 
-	float img_pso_pt_y 				= img_ref_pt_y + rel_pso_pt_y; 		// pso particle absolute coordinate
-	float img_pso_pt_x 				= img_ref_pt_x + rel_pso_pt_x; 		// pso particle absolute coordinate
+	double img_pso_pt_y 				= img_ref_pt_y + rel_pso_pt_y; 		// pso particle absolute coordinate
+	double img_pso_pt_x 				= img_ref_pt_x + rel_pso_pt_x; 		// pso particle absolute coordinate
 
-	float img_pso_pt_y_shift 		= img_pso_pt_y - ((subset_side_len - 1) / 2.0f); // shift to left_top_point, means (0,0) point in subset matrix(n,n)
-	float img_pso_pt_x_shift		= img_pso_pt_x - ((subset_side_len - 1) / 2.0f);
+	double img_pso_pt_y_shift 		= img_pso_pt_y - ((subset_side_len - 1) / 2.0f); // shift to left_top_point, means (0,0) point in subset matrix(n,n)
+	double img_pso_pt_x_shift		= img_pso_pt_x - ((subset_side_len - 1) / 2.0f);
 
 	if ((img_pso_pt_y_shift + subset_side_len + interp_safe_margin()) > img_height || (img_pso_pt_y_shift + interp_safe_margin()) < 0) {
 		SYS_DBG("shifted PSO particle image pos_y out of bound!\n");
@@ -136,17 +136,17 @@ float ZNCC_cost_function(struct PSO_context *ctx) {
 	
 	for (int row = 0; row < subset_side_len; row++) {
 		for (int col = 0; col < subset_side_len; col++) {
-			float img_pso_pt_y_shift_row = img_pso_pt_y_shift + row;
-			float img_pso_pt_x_shift_col = img_pso_pt_x_shift + col;
+			double img_pso_pt_y_shift_row = img_pso_pt_y_shift + row;
+			double img_pso_pt_x_shift_col = img_pso_pt_x_shift + col;
 			int pixel_idx = row * subset_side_len + col;
 			cur_subset_data[pixel_idx] = bilinear(cur_img_data, img_width, img_height,
 												img_pso_pt_x_shift_col, img_pso_pt_y_shift_row);
 		}
 	}
 
-	float sum = 0.0f, ref_sum_den = 0.0f, cur_sum_den = 0.0f;
-	float cur_subset_mean = get_subset_mean(&(zncc_ctx->cur_subset_info));
-	float ref_subset_mean = zncc_ctx->ref_subset_info.mean;
+	double sum = 0.0f, ref_sum_den = 0.0f, cur_sum_den = 0.0f;
+	double cur_subset_mean = get_subset_mean(&(zncc_ctx->cur_subset_info));
+	double ref_subset_mean = zncc_ctx->ref_subset_info.mean;
 	if (!ref_subset_data) {
 		SYS_DBG("Error: ref_subset_data is NULL!\n");
 		return NAN;
@@ -154,8 +154,8 @@ float ZNCC_cost_function(struct PSO_context *ctx) {
 
 	for (int row = 0; row < subset_side_len; row++) {
 		for (int col = 0; col < subset_side_len; col++) {
-			float mean_substrat_ref_subset = ((ref_subset_data[row * subset_side_len + col]) - ref_subset_mean);
-			float mean_substrat_cur_subset = ((cur_subset_data[row * subset_side_len + col]) - cur_subset_mean);
+			double mean_substrat_ref_subset = ((ref_subset_data[row * subset_side_len + col]) - ref_subset_mean);
+			double mean_substrat_cur_subset = ((cur_subset_data[row * subset_side_len + col]) - cur_subset_mean);
 			sum += mean_substrat_ref_subset * mean_substrat_cur_subset;
 			ref_sum_den += square(mean_substrat_ref_subset);
 			cur_sum_den += square(mean_substrat_cur_subset);
@@ -196,7 +196,7 @@ int SSD_init(struct SYS_INFO *info) {
 	return 0;
 }
 
-float SSD_cost_function(struct PSO_context *ctx) {
+double SSD_cost_function(struct PSO_context *ctx) {
 	return 0;
 }
